added wav file and new mathodds in GameField

+++ Hugo/Hugo/Helpers/Coord.cs
@@ -8,15 +8,15 @@
 
     public struct Coord
     {
-        public Coord(int x, int y)
+        public Coord(float x, float y)
             : this()
         {
             this.X = x;
             this.Y = y;
         }
 
-        public int X { get; set; }
+        public float X { get; set; }
 
-        public int Y { get; set; }
+        public float Y { get; set; }
     }
 }
